Fix cursor when drag-dropping in MML
